refactor: 공지페이지 컴포넌트 분리
